requirements.txt
attrs==22.1.0
Authlib==1.1.0
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
hubspot-api-client==6.0.0
idna==3.4
isodate==0.6.1
lxml==4.9.1
mysql-connector-python-rf==2.2.2
numpy==1.23.4
pandas==1.5.0
platformdirs==2.5.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.0
simple-salesforce==1.12.2
six==1.16.0
urllib3==1.26.12
zeep==4.1.0
